# Slightly faster Lexer::Elem#to_a