--- a/ext/ripper/lib/ripper/lexer.rb
+++ b/ext/ripper/lib/ripper/lexer.rb
@@ -95,7 +95,7 @@ class Ripper
 
       def to_a
         a = super
-        a.pop unless a.last
+        a.pop unless a.empty?
         a
       end
     end