@schneems 2021-11-05
Turns out I misunderstood what was going on, it is checking to see if message is nil, not if message exists. It will never be an empty array.

The updates to switch to a real class instead of a struct are still possible.